Export `DummyArkFrontend` from Ark
