+++ crates/ark/src/test/mod.rs
@@ -1,0 +1,5 @@
+pub mod dummy_frontend;
+pub mod test;
+
+pub use dummy_frontend::*;
+pub use test::*;
